move schema to root and log error if MAC not found in dancerTable
