+++ controller-server/schema/PinMapTable.ts
